@@ -22,28 +22,28 @@
     "@angular/platform-browser-dynamic": "~18.2.13",
     "@angular/router": "~18.2.13",
     "angular-oauth2-oidc": "^18.0.0",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.14.0"
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~18.2.20",
     "@angular/cli": "^18.2.20",
     "@angular/compiler-cli": "~18.2.13",
-    "@types/jasmine": "~3.6.0",
+    "@types/jasmine": "~5.1.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^13.13.4",
+    "@types/node": "^18.18.0",
     "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.8.0",
+    "jasmine-core": "~5.2.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~6.4.4",
-    "karma-chrome-launcher": "~3.1.0",
+    "karma-chrome-launcher": "~3.2.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.7.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~5.4.0"
+    "typescript": "~5.5.2"
   }
 }
