@@ -1,14 +1,36 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { RouterLink } from '@angular/router';
+import { provideHttpClient } from '@angular/common/http';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
+import { ActivatedRoute } from '@angular/router';
+import { of } from 'rxjs';
 
 import { MoviesComponent } from './movies.component';
 
 describe('MoviesComponent', () => {
+  const mockActivatedRoute = {
+    snapshot: {
+      data: { title: 'Mock Title' }
+    },
+    data: of({
+      movies: [
+        { title: 'Mock Movie' }
+      ]
+    })
+  };
+
   let component: MoviesComponent;
   let fixture: ComponentFixture<MoviesComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ MoviesComponent ]
+      declarations: [ MoviesComponent ],
+      imports: [ RouterLink ],
+      providers: [
+        provideHttpClient(),
+        provideHttpClientTesting(),
+        { provide: ActivatedRoute, useValue: mockActivatedRoute }
+      ]
     })
     .compileComponents();
   });
@@ -21,5 +43,6 @@ describe('MoviesComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+    expect(component.title).toBe('Mock Title');
   });
 });
