@@ -1,8 +1,10 @@
+import { ElementRef } from '@angular/core';
 import { HighlightDirective } from './highlight.directive';
 
 describe('HighlightDirective', () => {
   it('should create an instance', () => {
-    const directive = new HighlightDirective();
+    const mockElementRef = new ElementRef(document.createElement('div'));
+    const directive = new HighlightDirective(mockElementRef);
     expect(directive).toBeTruthy();
   });
 });
